+++ posix_spawn.c
@@ -197,7 +197,7 @@ loader:
 					nargv, __environ);
 	}
 
-	return __set_errno(EINVAL, -1);
+	return __set_errno(E2BIG, -1);
 
 exec_sh:
 	ret = __exec_sh(path, argv, interparg, buf, sizeof(buf));
@@ -231,5 +231,5 @@ exec_sh:
 					nargv, __environ);
 	}
 
-	return __set_errno(EINVAL, -1);
+	return __set_errno(E2BIG, -1);
 }